fix: update import paths after moving util files to utils folder
