+++ src/lib/ChatScreens/ChatBody.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
     import { isEqual } from "lodash"
     import { DBState } from 'src/ts/stores.svelte'
-    import { sleep } from "src/ts/util"
+    import { sleep } from "src/ts/utils/util"
     import { alertError } from "../../ts/alert.svelte"
     import { addMetadataToElement, getDistance, ParseMarkdown, postTranslationParse, trimMarkdown, type CbsConditions, type simpleCharacterArgument } from "../../ts/parser.svelte"
     import { getLLMCache, translateHTML } from "../../ts/translator/translator"
