+++ src/lib/ChatScreens/DefaultChatScreen.svelte
@@ -9,7 +9,7 @@
     import { DBState } from 'src/ts/stores.svelte';
     import { getCharImage } from "../../ts/characters.svelte";
     import { chatProcessStage, doingChat, sendChat } from "../../ts/process/index.svelte";
-    import { sleep } from "../../ts/util";
+    import { sleep } from "../../ts/utils/util";
     import { language } from "../../lang";
     import { isExpTranslator, translate } from "../../ts/translator/translator";
     import { alertError, alertNormal, alertWait, showHypaV2Alert } from "../../ts/alert.svelte";
