+++ src/lib/Playground/PlaygroundSyntax.svelte
@@ -2,7 +2,7 @@
     import TextAreaInput from "../UI/GUI/TextAreaInput.svelte";
     import { risuChatParser } from 'src/ts/parser.svelte';
     import { language } from 'src/lang';
-    import { sleep } from 'src/ts/util';
+    import { sleep } from 'src/ts/utils/util';
     let input = $state("");
     let output = $state("");
     const onInput = async () => {
