+++ src/lib/Playground/ToolConversion.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
     import { language } from "src/lang";
     import Button from "../UI/GUI/Button.svelte";
-    import { openFilePicker } from "src/ts/util";
+    import { openFilePicker } from "src/ts/utils/util";
     import { detectPromptJSONType, promptConvertion } from "src/ts/process/prompt";
 
     let files: { name: string, content: string, type:string }[] = $state([])
