+++ src/lib/SideBars/Scripts/RegexList.svelte
@@ -2,7 +2,7 @@
     import type { customscript } from "src/ts/storage/types/character";
     import RegexData from "./RegexData.svelte";
     import Sortable from "sortablejs";
-    import { sleep, sortableOptions } from "src/ts/util";
+    import { sleep, sortableOptions } from "src/ts/utils/util";
     import { onDestroy, onMount } from "svelte";
   import { DownloadIcon, HardDriveUploadIcon, PlusIcon } from "@lucide/svelte";
   import { exportRegex, importRegex } from "src/ts/process/scripts";
