+++ src/lib/UI/GUI/TextAreaInput.svelte
@@ -100,7 +100,7 @@
 <script lang="ts">
     import { textAreaSize, textAreaTextSize } from 'src/ts/gui/guisize'
     import { highlighter, getNewHighlightId, removeHighlight, AllCBS } from 'src/ts/gui/highlight'
-    import { sleep } from 'src/ts/util';
+    import { sleep } from 'src/ts/utils/util';
     import { onDestroy, onMount } from 'svelte';
   import { disableHighlight } from 'src/ts/stores.svelte';
   import { isMobile } from 'src/ts/platform'
