+++ src/lib/VisualNovel/VisualNovelChat.svelte
@@ -3,7 +3,7 @@
     import { DBState } from 'src/ts/stores.svelte';
     import { getFileSrc } from "src/ts/globalApi.svelte";
     import { selectedCharID } from "src/ts/stores.svelte";
-    import { sleep } from "src/ts/util";
+    import { sleep } from "src/ts/utils/util";
     import { onDestroy, onMount } from "svelte";
 
     const style:number = 1
